backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Veritabanı
sqlalchemy==2.0.30
psycopg2-binary==2.9.10
alembic==1.13.1

# Veri & ML
pandas==2.2.2
numpy==2.2.0
scikit-learn==1.5.2
joblib==1.4.2

# Validasyon & Ayarlar
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS & Yardımcılar
python-multipart==0.0.9
httpx==0.27.0